+++ compiler/example/test-utils/output-group-name-with-hyphen-datamodel/_generated_base_structure/nexus-client/errors.go
