+++ compiler/example/test-utils/output-group-name-with-hyphen-datamodel/crd_generated/api_names.sh
