+++ compiler/example/test-utils/output-group-name-with-hyphen-datamodel/crd_generated/crds/config_config.yaml
