+++ compiler/example/test-utils/output-group-name-with-hyphen-datamodel/crd_generated/crds/project_project.yaml
