+++ compiler/example/test-utils/output-group-name-with-hyphen-datamodel/crd_generated/crds/root_root.yaml
