+++ compiler/example/test-utils/output-group-name-with-hyphen-datamodel/crd_generated/nexus-gql/gqlgen.yml
