+++ compiler/example/test-utils/output-group-name-with-hyphen-datamodel/go.mod
