+++ compiler/pkg/generator/generator_test.go
@@ -198,8 +198,8 @@ var _ = Describe("Template renderers tests", func() {
 	It("should handle hyphen in group-name", func() {
 		datamodelPath := "../../example/test-utils/group-name-with-hyphen-datamodel"
 		groupName := "tsm-tanzu.vmware.com"
-		crdModulePath := "../../example/test-utils/test-output/crd_generated/"
-		outputDir := "../../example/test-utils/test-output/crd_generated"
+		crdModulePath := "../../example/test-utils/output-group-name-with-hyphen-datamodel/crd_generated/"
+		outputDir := "../../example/test-utils/output-group-name-with-hyphen-datamodel/crd_generated"
 
 		pkgs := parser.ParseDSLPkg(datamodelPath)
 		graphlqQueries := parser.ParseGraphqlQuerySpecs(pkgs)
